Pass a VALUE_PAIR** to valuepairs callback function
It seems I have mis-read the API for userparse - it needs a pointer to
the first list item, i.e. &request->vps -- Adding other items worked
with the old version, but adding pairs might leak memory. This should be
fixed now.

WARNING: Even though this does not change the ABI, it *does* change the
contract for the callback function: A callback function must now receive
the argument and cast it to (VALUE_PAIR **).